requirements.txt
Faker==8.10.1
pymongo==3.11.4
python-dateutil==2.8.2
PyYAML==5.4.1
six==1.16.0
text-unidecode==1.3
